requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.0.1
django-ckeditor==6.4.0
django-js-asset==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
Pillow==9.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
